詳細入力フォームのリファクタリング
